Добавили загрузку объявлений с сервера. Добавили отправку обявления на сервер
