+++ js/map.js
@@ -1,7 +1,6 @@
 /* global L:readonly */
 import {activatePage, addressBlock} from './page-state.js'
 import {getTemplateMarkup} from './ad-markup-generator.js';
-// import { getSimilarAds } from './similar-ads.js';
 
 //Центр Токио
 const TOKIO_CENTER = {
@@ -19,7 +18,7 @@ const moveMarker = (marker) => {
 
 
 // Добавляем главный маркер на карту
-const getMainMarker = () => {
+const createMainMarker = () => {
   const mainPinIcon = L.icon({
     iconUrl: 'img/main-pin.svg',
     iconSize: [52, 52],
@@ -40,7 +39,7 @@ const getMainMarker = () => {
 }
 
 //Добавляем точки объявлений
-const getAdvertisingMarkers = (adsArray) => {
+const createAdvertisingMarkers = (adsArray) => {
   if (adsArray) {
     adsArray.forEach((ad) => {
       const {location} = ad
@@ -70,7 +69,6 @@ const getAdvertisingMarkers = (adsArray) => {
 //Отображение карты
 const showMap = (adsData) => {
   map.on('load', () => {
-    status === true
     activatePage()
   })
     .setView({
@@ -84,8 +82,8 @@ const showMap = (adsData) => {
       attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors',
     },
   ).addTo(map);
-  getMainMarker();
-  getAdvertisingMarkers(adsData)
+  createMainMarker();
+  createAdvertisingMarkers(adsData)
 }
   
-export {showMap, TOKIO_CENTER, getAdvertisingMarkers, getMainMarker, moveMarker}
+export {showMap, TOKIO_CENTER, createAdvertisingMarkers, createMainMarker, moveMarker}
