+++ js/page-state.js
@@ -1,4 +1,5 @@
-import {TOKIO_CENTER} from './map.js'
+import {TOKIO_CENTER} from './map.js';
+import { userForm } from './event-listeners.js'; 
 
 const adForm = document.querySelector('.ad-form');
 const mapFilters = document.querySelector('.map__filters');
@@ -36,4 +37,12 @@ const activatePage = (status) => {
   }
 }
 
-export {activatePage, deactivatePage, addressBlock}
+
+// Функция очистки полей и установки значений по умолчанию в поле адрес
+const cleanFields = () => {
+  userForm.reset();
+  addressBlock.value = `${TOKIO_CENTER.x.toFixed(5)},${TOKIO_CENTER.y.toFixed(5)}`
+}
+
+
+export {activatePage, deactivatePage, addressBlock, cleanFields}
